Mobile control switching works using user-agent string parsing, restructured A-Frame component imports
