+++ src/index.js
@@ -1,7 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import "./utils/AframeReg";
+import "./utils/AFramePhysics";
+import "./utils/WASDPlusControls";
 import "bootstrap/dist/css/bootstrap.css";
 import "./css/App.css";
